requirements.txt
pycryptodome==3.22.0
sympy==1.13.3
ttkbootstrap==1.10.1
ttkbootstrap